+++ src/geoxgb/__init__.py
@@ -34,4 +34,4 @@ def load_model(path):
     return joblib.load(path)
 
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"